Reduce terminals to one lock.
The idea behind separate terminal input and output locks was to allow
slave to master I/O to happen in parallel with master to slave I/O. In
practice, this doesn't seem to be a common enough case to support given
that it adds complexity to the locking scheme. Also, because a master's
writes echo, the equivalent locking of a slave write is necessary, which
kills the parallelism anyway. Simplify life and move to one lock.