Avoid a badly timed PLT binding during OsExitThread.
Taking a PLT fault on OsExitThread is bad news. Detached threads clean up
the OS thread state just before exiting. With the image lock switched to a
read-write lock, acquiring that lock to resolve the PLT entry references
the current thread.

Call OsExitThread once in a special way that does nothing when threading is
initialized to make sure it's warmed up by the time a thread really needs
to exit.