Change user mode exception forwarding.
Previously, if a user mode thread received an unhandled signal and it was
not being debugged, then the exception would be passed on to the kernel.
This did not mix well with the behavior desired by sigwaitinfo, where a
thread waits for a particular set of signals, but does not necessarily
handle them. The kernel was breaking in even though the user mode thread
was expecting an unhandled signal.

This change makes it so only the unhandled and unexpected signals get
passed onto the kernel.