Fix pointer passed to 802.11 timeout work item.
The 802.11 state timeout worker was expecting a pointer to a NET80211_LINK,
but was receiving PNET_LINK. This change fixes this and brings the timeout
DPC to expect the same.

The change also cleans up tearing down a NET80211_LINK. Since the timeout
worker always attempts to acquire the NET80211_LINK's lock, the destroy
routine needs to make sure that the worker is flushed and destroyed before
destroying the lock.