Allow 802.11 scans for the active BSS.
A 802.11 network join request may be issued for the active BSS. Go through
all the steps to make sure it is still within range, but if the BSS is
still in range, do not tear down the state. Simply reconnect to it. Tearing
down the state was causing reference count issues, exposing this bug.