Add some debug instrumentation for signal failure.
Somewhere inside of signal handling on ARM there exists a corruption.
Everything actually looks pretty normal, but R0 gets corrupted somewhere.
It seems to get restored to its pre-signal value, even though it's in the
middle of trying to handle that signal. I'm not sure if it's the result
of a nested signal, so I'm saving the last signal restore context so
that I can examine it.