Remove O_BINARY and O_TEXT flags.
These flags are not necessary and not honored on Minoca OS. They were only
present because of build tools built to run on Minoca and Windows. For
those tools, this change trades the use of open with O_BINARY for fopen
with the 'b' mode character.

It is necessary to remove them because third party packages (e.g. Ruby)
were assuming the presence of O_BINARY meant the presence Win32's setmode.

Note that swiss still defines O_BINARY in swlibos.h so it can use open and
maintain compatability with Windows.