Fix reference accounting in PsSysSendSignal.
PspSignalProcess got created as a helper routine to PsSysSendSignal, but
carried over a release reference in an uncommon failure case. This change
removes it from PspSignalProcess.

PsSysSendSignal was not releasing a reference on the process taken by
PspGetProcessById if the found process was the kernel process. Add the
release reference.