Support MmZeroIoBuffer to user mode buffers.
The node tests were using /dev/zero in a way that caused user mode IO
buffers to end up all the way down in the driver. MmZeroIoBuffer was
asserting that it shouldn't be seeing user mode buffers. Fix it up to
support those, since it seems to be useful to drivers.