Modify system call handler function signature.
This change reduces the number of parameters passed to the
PSYSTEM_CALL_ROUTINE. The system call number was only used in asserts and
is no longer necessary. The trap frame is now stored in the KTHREAD
structure. The result size was only used by PsSysExecuteImage and is
replaced by including a system call parameter copy out size in the system
call table entry structure.

In order to allow PsSysExecuteImage to return an error status without ever
copying the parameters out to user mode, which would cause an access
violation if the failure occurred after a certain point, this change also
allows system calls to return a value via a register (r0 on ARM and eax on
x86). This requires modifying the system call assembly handlers to not
restore the return register from the trap frame and to not scrub it when
only a partial trap frame was saved.

In turn, this complicates the system calls where a full trap frame needs
to be restored, including the return register. As a result,
PsSysRestoreContext and PsSysExecuteImage (when successful) need to return
the architecture's return register value which would normally have just
been restored by the trap frame.