Fix for ARM signal corruption issue: clear IT state.
I believe this will fix the signal corruption issue we've been seeing on
ARM. From examining the registers, it looked like everything went fine, but
the initial 'mov r0, sp' never happened. I think this is because we
applied the signal in the middle of an If-Then Thumb conditional section
that was not being executed. So it really was that the next one to three
instructions wouldn't get executed.

The solution is simple: reset CPSR when applying a signal.