Fix for paging related crash on ARM.
ARM was crashing again because it was doing MmpPrepareForPageFileRead on a
section that had no page file space (because it was never marked
WAS_WRITABLE). It probably got marked dirty because the parent had
WAS_WRITABLE set, and the parent called isolate. Isolate will mark the
child dirty as well.

The child should inherit WAS_WRITABLE from the parent, which it wasn't
previously doing. This is needed for exactly the case above: if a parent
isolates and disappears, the child needs to know that it has potentially
dirty data mapped and cannot just throw the page away.