Some support for a C interface in Chalk.
This change adds support for foreign data objects, and some external APIs
for manipulating variables, functions, classes, and calls.

I realized though that I'm going through a lot of gyrations to support
indirect calls to classes. If I used the construct opcode and then a
regular call to __init, a lot of the code would probably collapse
together. So I may give that a try before coming back to finishing and
testing this C API.