Return from ObWait immediately on timeout of 0.
If ObWait is called with a timeout of zero, there's no need to actually go
down for the wait. The timer queue code does not check to see if the
deadline is already expired, so it may take a full clock tick before that
thread gets to run again. A glib test was failing to complete because it
took so long to do these waits.

I'm curious to see if this leads to any perf gains, as it eliminates a
bunch of needless blocked time.