Add support to restart system calls after signals.
This change adds support for the SA_RESTART flag used with sigaction. Right
now the only system call that supports restarting is
PsSysWaitForChildProcess. Support for the others will be added shortly.

A system call restart is viable if the system call returns a new KSTATUS -
STATUS_RESTART_SYSTEM_CALL. A system call looking to dispatch a signal
must pass its return value, paramters, and system call number. The signal
application code will evaluate the return value and system call type to
decide whether or not a restart is viable. If so, it will set a flag in
the signal context to alert user mode of the possibility of a restart and
save the system call number and paramters in the signal context. If the user
mode signal handler OKs the restart, it leaves the flag; if it
does not allow the restart, it will clear the flag. Then on restore, if
the flag is still set, the main system call handler is called again with
the saved parameters.

This change also makes sure that a system calls return value is preserved
across a handled signal, even if a restart isn't in play. This will allow
the system calls to return a status or parameter via a register.