Reshaped SIGNAL_CONTEXT in preparation for ucontext.
The SIGNAL_CONTEXT structure now contains a signal mask, all general
registers, and potentially the floating point registers. This is in line
with the requirements of the ucontext structure. Saving all the registers
may also make it easier for ptrace to manipulate registers in the target
once the user mode debugger stuff is reworked.

The trouble is getting the non-volatile registers during a system call.
The check for signals is going to have to be done in the system call
assembly, once all non-volatile registers have been restored by C code.
I'll work on this once the other signal code is in.