Set CR3 to the kernel TSS page directory on NMI.
When an NMI trap occurs, CR3 gets loaded with the value stored in the NMI
TSS. This is initialized to the kernel's page directory on processor bring
up and does not change. As a result, when the debugger breaks in and sends
NMI's to freeze the other processors, those other processors are stuck
running on the kernel's page tables and cannot access user mode addresses.

This change sets CR3 to the page directory in the processor's kernel TSS
when dispatching an NMI. This page directory should be the page directory of
the interrupted process as the kernel TSS is updated on context swap. This
allows access to user mode addresses.