Revert forced signal lock reacquisition workaround.
PspQueueSignal no longer sends a kill signal if the forced signal is found to
be blocked. The kill occurs naturally by unblocking the forced signal. The
prior call to send a kill signal causes locking issues, as the signal
sending routines tried to reacquire the process lock, already held by the
caller of PspQueueSignal. Now that this is no longer needed, revert the
workarounds that were put in place. They weren't very pretty.