Set the pending signal bitmask for all signals.
Prior to this change, the pending signal bitmasks in the thread and process
were not set if the signal came with a queue entry. This change makes it
so that all signals set a bit in the bitmask. As a result, dequeue and wait
need to be more intelligent about unsetting the bitmask - they can no longer
do it if a second signal of the same type is in the queue. The goal of this
change is to make checking to see if there is a pending signal as cheap as
looking at the bitmasks.