Remove ThreadSignalPendingStateUnknown.
Now that all signals set a bit in the pending mask, it is much cheapr to
keep a thread's pending signal state accurate. It used to be set to 'unknown'
to defer the decision making process to the next attempt to dequeue a
signal. But, as this signal rework is heading towards saving full trap frames
on the way out of system calls if there is a signal pending, this change
will help avoid the extra work the next time around if it isn't necessary.