Move syscall signal dispatching to assembly.
This change moves the standard system call handler's check for pending
signals on a thread to assembly. This allows the assembly to build up a
full trap frame for the ucontext even if the system call did not originally
save a full trap frame.

Note that PsSysWaitForChildProcess, IoSysPoll, and PsSysSuspendExecution
still dispatch signals. Those will be fixed once system call restarts are
updated, which is the next change.