Remove signal application from PsSysSuspendExecution.
This changes PsSysSuspendExecution to only dequeue signals if they are not
going to get applied (.e.g they are part of the mask supplied to sigwait).
For the other cases (pause, sigsuspend) it simply changes the blocked
signal mask and then suspends itself. If the suspend every wakes up for a
real signal (not a child), then the system call exits to attempt to
dispatch the signal. It may not succeed if another thread intercepts the
signal. This is a bug for pause/sigsuspend in multi-threaded applications.
The next change will fix this with an improved system call restart
mechanism."