Make sure device idle timer is queued correctly.
A race between PmDeviceReleaseReference and PmpDeviceIdleWorker could
result in the idle timer not being queued by either party.

This change refactors trying to win the compare exchange and queuing the
timer into a helper function. The timer queued variable is then
unconditionally cleared by the idle worker, and the helper function is
used to potentially requeue the timer.

Also ignore failures to queue the idle timer, since it's not critical and
no one really seems prepared to do anything should that happen.