Fixes for power transition machinery.
1. PmpDeviceAddReference could return STATUS_NOT_READY during a badly timed
idle transition. If it waited on the event which was not yet signaled, but
then the state changed to Transitioning, it would fail the add reference with
not ready even though it really was okay. Do a check for active at the
beginning before waiting on the event to prevent this situation.
Unsignaling of the ActiveEvent now happens in QueuePowerTransition, before
changing the state, so that add reference can't fall through.

2. In PmpDeviceResume, only send a resume IRP if the state is not
transitioning out of active. Before this change we may have been sending
resume requests corresponding to idle/suspend requests that never actually
got sent.

3. PmpDeviceIdle had holes where it would bail out early and leave the
state stuck at Transitioning. This might eventually fix itself with a few
add/release references. This change adds some code so that if it's going
to bail out, it may have to cancel the transition and move the state back
to active itself. This may occur if an add reference went through just
before the idle transition was queued.