Add system call restart support.
This change allows system calls to restart if they return one of a few
special status codes. The restart may occur if no signal is applied during
dispatch or if a signal handler allows restarts after it completes. The
restart mechanism is to back up the instruction pointer and then restore
the system call parameters in the trap frame so that the return to user mode
just calls right back into the kernel.