Make ArIsTrapFrameComplete accurate again.
When a fast system call upgrades to save a full trap frame, update the fake
values that signify an incomplete trap frame (-1 exception CPSR on ARM and
setting CS to user DS on x86). This makes ArIsTrapFrameComplete accurate at
all times for checking how much of a trap frame can be copied to user mode.

As a result of this change, x86 needs a different way to check to see if
a system call came from sysenter or int 0x2F in order to know how far to
back up the EIP on a restart. x86 will now read the instruction stream
looking for int 0x2F in order to determine if it was a normal or sysenter
system call.