Dispatch signals only if ThreadSignalPending is set.
Now that the 'unknown' thread state has been removed, signals only ever need
to be dequeued when the thread state is exactly ThreadSignalPending. This
change updates the system call assembly routines and the dequeue routine to
bail if a thread's signal pending state is not ThreadSignalPending. The
ThreadChildSignalPending state really just means that a child is available
for wait, not that a child signal was actually sent.

This change also introduces a fix for ARMv6 in the system call assembly.
The cbz/cbnz instructions are not available on ARMv6k.