Add restarts to pause and sigsuspend if no signal.
Pause and sigsuspend should only return if a signal is actually dispatched.
This change adds functionality to PsSysSuspendExecution to use the
STATUS_RESTART_NO_SIGNAL status. This status signifies that a system call
should be restarted if no signal is applied by dispatch. If a signal is
actually applied, the status will be converted to STATUS_INTERRUPTED in
the saved trap frame.