Preserve I/O size in case of restarts.
For both regular and socket I/O, the system call parameters size was used
to pass the number of bytes requested for the I/O and to return the number
of bytes actually processed. The I/O routines were setting the size to 0 as
soon as it was read into a local. This is problematic if the I/O needs to
restart due to a signal. The restarted system call would think the request
was for 0 bytes. This change preserves the size field and introductes a
bytes completed field to track that actual work done.