Fix minor pipe read regression.
When pipes are being read with no writers, it turns the reads into
non-blocking reads. After IoReadStreamBuffer returns, the pipe code may
check the status code and writer count and turn it into EOF. A recent
change to the status code returned in this case left the pipe code
checking for the wrong status code. So the read would fail instead of
returning 0. This causes Tcl to fail to build.