Convert IoSysPoll to use LONG descriptor count.
This change enables IoSysPoll to not have to copy out any parameters, as
the positive selected descriptor count can overlap with the negative KSTATUS
codes and be returned via a register.

The OsPoll interface, however, does not change. It would not make sense to
pass a negative descriptor count to OsPoll. So, OsPoll simply returns an
error if more than MAX_LONG descriptors are requested and then converts the
supplied ULONG to a LONG in the system call parameters.

This is all OK for the C Library because POSIX actually limits the unsigned
nfds_t argument to poll/ppoll to less than {OPEN_MAX}, which should always
be less than or equal to MAX_LONG.