Convert IoSysPerformIo to use an INTN size.
This change enables IoSysPerformIo and IoSysPerformVectoredIo to not have to
copy out any parameters, as the positive bytes completed can overlap with
the negative KSTATUS codes and be returned via a register.

The OsPerformIo and OsPerformVectoredIo interfaces, however, remain
unchanged. It would never make sense to request a negative byte count for
an I/O operation. So, the OS APIs simply truncate the request down to
MAX_INTN and pass an INTN to the kernel. The kernel still checks against
negative sizes in case someone malicious circumvents the OS APIs.

This is all OK because doing I/O larger than 2GB at a time is ridiculous
and POSIX even allows truncating the size_t supplied to reads and writes
down to SSIZE_MAX.