SD Fixes for Veyron.
When an error occurs, RK32 SD now shuts the power completely off to the
card using the RK808 regulator, so that it can reinitialize it
successfully. This allows us to make progress even with occasional SD
errors.

It also brought to light an issue in the RK32 SD controller, which was I/O
that was bigger than 1 page worth of DMA descriptors would get messed up,
because the SendBlockCount command sent the full size, but DMA only
completed the partial size.

Finally, add some extra prints to try to understand an occasional
BeagleBone Black SD error that seems to crop up.