Fix cache-backed section dirty bitmap accounting.
Cache-backed image sections depend on the dirty bitmap to know whether or
not to free a physical page when unmapping the section for destroy or
truncate. Truncate was unsetting the dirty bitmap before reading it to
determine if the page should be freed on unmap. As a result, pages were
leaking.

This was caught because page out found a PAGING_ENTRY attached to a
destroyed image section and hit an assert.

In order to make sure the dirty bitmap is accurate for cache-backed image
sections, this change includes more than just the truncate fix. It makes
sure that isolation (for clip) does not create a private page for an image
section that was never writable. It also sets the 'was writable' image
section flag if a read-only section becomes writable. Page out, isolation,
etc depend on the was-writable flag and not the writable flag.

Lastly, it makes sure that cache-backed sections do not set a PAGING_ENTRY
when paging in from the page cache. Truncate could cause an image section
to go from dirty to clean after a PAGING_ENTRY has been allocated.
Previously this would have smashed the page cache entry with a
PAGING_ENTRY in the physical page array. Bad news.