Update dev@minocacorp.com to minoca-dev@googlegroups.com.
My original plan was to have one of the members of dev@minocacorp.com be
minoca-dev@googlegroups.com. But Google doesn't love it when it gets mail
from dev@minocacorp.com with an envelope From of rando@somedomain.com,
because the SPF records of somedomain.com indicate that
mail.minocacorp.com is not on the list of senders for somedomain.com.
While Google doesn't seem to fail it outright, it calls it a 'softfail'.
I'm not confident that won't turn into a hard fail later, so I'm
abandoning this scheme in favor of just sending people towards
minoca-dev@googlegroups.com.