Relax page in crashes from user mode.
MmpPageIn can encounter non-fatal errors in addition to end-of-file that
should not result in a crash. For instance, if an image section maps a file
on a USB drive that is removed unsafely, access to that image section after
the removal should result in an access violation rather than a crash.