Add .identifier support to Chalk.
With this change a.b compiles to a.__get("b"), and a.b = c compiles to
a.__set("b", c). There is a slight ambiguity between the indirect call of
x.myprop() and the method call of x.myfunction(). Right now the resolution
is to assume a.b() is a method call. If the caller wants a.__get("b")()
they can either say that directly or use parentheses for (a.b)(). It's not
ideal, but properties seem necessary enough, and using a new operator like
-> seemed too distracting just to handle this case.

Also fixed some bugs with precedence in yygen, and converted the binary
operators to use precedence rather than cascading rules.

Also fixed an infinite recursion if the unhandled exception handler took an
exception, and fixed an invalid stack layout for slice assign.

__sliceAssign and __set are not ideal. The rvalue gets evaluated before
the lvalue receiver, which means the stack is in the wrong order. The
current solution is to reload the rvalue from the stack after the receiver
and key are loaded. This requires an extra load and pop, which I don't
love. I was thinking about other mechanisms where the CK_COMPILER.Assign
turns into an expression node pointer, that may get evaluated at the right
time during assignment (ie load receiver, load key, then evaluate
rvalue). But things get sticky with preincrement, as in that case the
rvalue is a computed value on the stack, so you can't stick a single node
as a pointer anywhere.

Another solution might be a Rotate3 stack opcode, but that seems cheesy.