Add TCP timer reference count indirection.
The Raspberry Pi 2 crashed with a double queue of the TCP timer. Tracking
that down is tough, and I can currently only really explain it with an
extra inappropriate release at the wrong time.

Keep a timer reference count in the socket, and only do the atomic
operations of the global timer reference count if the socket really needs
to. This should save on atomic operations, and also make the timer
reference count a bit clearer, as each socket should only ever have at
most one reference on the global TCP timer.