Clean evicted page cache entries correctly.
Page cache eviction was cleaning a page cache entry and then checking its
reference count to see if it could be destroyed immediately (a reference
count of 0). The problem is that the cache entry may (1) be cleaned by
eviction, (2) marked dirty by another thread, (3) be released by another
thread, (4) seen with a reference count of 0 by eviction. The entry would
end up on the destroy list and be dirty. This would mess up the accounting
and hit asserts. This change fixes that.

The change also skips some list juggling for page cache entries that have
already been evicted. It tries to keep them to the removal list.

Lastly, it makes sure to only put a page cache entry on a local destroy
list if it has the correct number of references while the global list lock
is held. Page cache entries can get new references not only when the file
object lock is held, but also when the list lock is held.