Only create a page cache entry if lookup fails.
IopCreateOrLookupPageCacheEntry would create a page cache entry before
checking to see if a page cache entry already existed at the given offset.
It would then delete the newly created entry if the lookup succeeded. This
was a remnant of past locking schemes. Under the current scheme, with the
file object lock held exclusively, a page cache entry only needs to be
created if the lookup fails.