Add bundle module to Chalk.
Chalk can now create a self contained executable that consists of the
modules it needs to run a complete application. This should make Chalk
apps more distributable.

Currently chalk.dll/libchalk.so is still needed, so it's a two-binary
solution instead of a single binary solution. The trouble is that on
Windows DLLs need to import against something. So they need chalk.dll.