Fixed file object / image section lock inversion.
The lock order between file objects and image sections dictates that the
file object lock is acquired first. IoMarkPageCacheEntryDirty caused a
violation in this order when unmapping shared image sections. The fix is
to introduce the PAGE_CACHE_ENTRY_FLAG_DIRTY_PENDING bit. This almost has
the same effect as PAGE_CACHE_ENTRY_FLAG_DIRTY except it does not change
the accounting numbers (which require the file object lock to update). As
a result, IoMarkPageCacheEntryDirty can just set the dirty pending flag and
not need the file object lock to update the accounting numbers.

The downside is that the page cache is temporarily dirtier than the
accounting numbers indicate, but this was always the case with shared image
sections. The page cache can only get notified about their dirtiness when
they get unmapped. This change just delays that a little further.