Move setup over to the new Chalk interpreter.
This change activates the new Chalk interpreter in the setup program. It
consolidates the platform recipes into one script, and adds an internal
msetup module to pass the configuration parameters like platform and
architecture.

This change also fixes several bugs in the new Chalk interpreter, and
enables storing a context pointer inside the CK_VM object.

Since the new Chalk interpreter uses const, this change also bleeds consts
out throughout setup. In fact, there are consts that bled out throughout
the repository, but those are coming in a future change in a couple of
minutes.