Convert page counts from ULONGLONG to UINTN.
The physical page allocator deals with page count requests, rather than byte
requests. As a result, it does not need to use ULONGLONG's to handle 32-bit
PAE; the maximum page index on 32-bit PAE only uses 24-bits. A UINTN allows
32-bit architectures a maximum of 16 terabytes of physical memory, which
seems like more than enough for the future of 32-bit.

This change allows the paging spin lock to be removed.