Build fix for build Chalk using Mac Homebrew.
The clang-pretending-to-be-GCC that comes with xcode seems to pass
an include directive to the assembler to include the directory that
the source file being assembled is in. The GCC from Homebrew,
however, does not seem to do that. This was causing .incbin failures
when building Chalk for the build environment. Explicitly include
$(SRCDIR)/.. for the assembler during this step.