ACPI: Fix for lack of keyboard input in VMWare 12.
The _HID method for one of the devices behind the PCI-to-ISA
device was returning a FieldUnit object. _HID expects only
an integer or string, so it was failing to enumerate any
children behind the ISA bus, including the keyboard.

Since none of the top level ACPI methods seem to specify a
FieldUnit as a possible return value, the fix is to read from
the field unit if a field unit is returned to a top level
method (ie a value returned from AcpiExecuteMethod).