Treat all ACPI ASL stores to Zero as no-ops.
ASL Operators that store a result to an optional argument use Zero to
indicate that the store should not be performed. The ACPI driver was
previously attempting the store. It was succeeding when the source was an
integer but finally reported an error when it tried to store a Package
into an Integer (zero).

This change also makes Zero, One, and Ones globals so they don't get
created all the time.