Remove armv6 from the symbols zip.
It takes a week for the Raspberry Pi 1 to build all packages in the
repository. This makes it very difficult to cut a release because nightly
builds basically have to stop during that week, otherwise the armv7/x86
symbols are a different build from armv6.

Eventually we'll need a new build system that can farm out packages to
many machines. With multiple Raspberry Pi 1s working in parallel, maybe we
can consider re-adding armv6 to the symbols distributable.