Fix for missing prompt on rel builds.
When the first user mode program was coming up (libminocaos.so), it starts
by relocating itself. ImpElfGatherExportInformation was reaching through a
global ImImportTable to get the resolver function to store in the GOT. The
problem is reaching through that global requires relocations to have been
completed, which they have not. So that function reaches through the
unrelocated address to get at the global, which is an invalid address.
What's more, ImInitialize hasn't been called yet, so there's nothing in
that pointer anyway.

The fix is to pass the resolver function down directly, so that we don't
have to reach through a global to get it.

I have no idea how this was working on dbg builds. If we had gotten lucky
with the unrelocated address and its dereference, then I suppose it may
have worked since libminocaos.so has no unresolved PLT entries.