Fix AML Store and CopyObject to LocalX objects.
In AML, LocalX objects get reassigned on writes without any type
conversions. Any previously existing object at LocalX is deleted
(dereferenced) and a copy of the object being written is left in its place
at LocalX.

Because LocalX objects are evaluated before Store or CopyObject writes to
them, the write's destination is passed as the existing object in LocalX,
without any indication that it came from a local. This change records the
index of the last local evaluated so that writes can detect that the
destination is in fact an object that came from a LocalX. The last local
index must be cleared at the end of all non-local statements so it does
not accidentally get interpreted later on.