Remove -nt test for potentially non-existant files.
The postbuild step for many apps was trying to be nice and not copy the
file if it hadn't been updated. On other OSes, the test -nt (newer than)
was failing if the file didn't exist. So the copies never happened, and
the final image was missing important user mode programs, and never
getting to the prompt.