Toss mouse data in i8042 interrupt handler.
While running Minoca from QEMU on Linux, pressing Ctrl-Alt to exit QEMU's
mouse grab was causing mysterious '7' keystrokes the appear. This was
caused by QEMU sending mouse data to the i8042 device. It is easily
reproduced by clicking to focus on the QEMU window, moving the mouse
around, and then pressing Ctrl-Alt. As Minoca's i8042 driver does not yet
handle mice, it was interpreting the mouse data as keystrokes - generating
the string of 7's. The fix is to ignore mouse data in the i8042 interrupt
handler.