Improve debugger UI preference tracking.
The debugger can still, somehow, save window state to the prefs file with
a height and width of 0. This change protects against that by never saving
or restoring preferences that would result in a window with zero height or
width. It also makes some improvements to tracking the main pane and
profiler dividing lines.