Fix glob pattern matching regression.
Static analysis indicated that the GLOB_META_* macros should be casted to
CHARs. They were being compared to CHARs throughout the glob code. This
change broke ClpGlobMatch, causing an entire switch statement to be
compiled out due to integer promotion differences between the GLOB_META_*
CHAR macros and a CHAR being AND'd with 0xFF.

The fix is to remove the GLOB_META_MASK (0xFF). It is not needed as the
GLOBL_META_* macros are only ever being compared to CHARs.