Bug fixes for booting under Hyper-V.
ATA had a bug where after it tried the non-existent second drive on the
cable, it left that as the selected drive.

i8042 seemed to have a bug where if there were keystrokes in the buffer
before the interrupt was connected, and interrupt would never come in. The
solution is to clear out the buffer after the interrupt is connected.