Fix TCP socket timer reference accounting.
On entering TcpStateClosed, a socket's timer reference count was being
forced to 1 and then released. This caused the socket data buffer free
routine to assert when releasing the reference the outgoing segments list
thought it had on the timer. This change reverts back to more precise
accounting by looking at the state the socket was in before being set to
closed.

It also asserts that the timer reference is 0 on destroy so it can
catch whatever bug drove the decision to force the reference count to 1.