Fix TCP window field in SYN packets.
According the RFC 7323, "The window field in a segment where the SYN bit
is set MUST NOT be scaled." Minoca's network stack was sending SYN packets
with a scaled window size of 256, thinking the other end would shift it by
8 because of the window scaling advertisement. But it was being
interpreted as a 256 byte window, causing the first packet sent by the
other end to be capped at 256 bytes. Of course, the requirement makes
sense because it is not yet known whether or not the other end supports
window scaling. The fix is to temporarily change the receive window scale
and size when filling out the TCP header for SYN packets.