Fix RTL8139 NIC support for QEMU.
QEMU offers the RTL8139C+ as a NIC. It, however, does not follow the
specification with regard to the basic mode control register's reset bit.
It is supposed to be self-clearing, but inspection of the QEMU source
shows that it is never cleared when set. As a result, PHY initialization
was timing out. The workaround is to simply allow the device to contiue
initialization even if the reset bit does not clear before the 1 second
timeout.