Override I/O timeouts when bytes are completed.
The perform I/O system calls used to always return a status and the number
of bytes completed. So, even in failure cases, the bytes read or written
could be observed. The C library was using this behavior on read and write
to convert STATUS_TIMEOUT to success if there was a non-zero byte count.

The perf change to return the bytes completed to user mode via a register
(overlapping it with the negative status space) broke this conversion
behavior. STATUS_TIMEOUT was being returned, but the completed byte count
was not making it to user mode. The fix is to have the kernel always convert
STATUS_TIMEOUT to the positive integer byte count if bytes were read/written
before the timeout occurred.