Fix TCP timer race condition.
The queuing of the TCP timer was not properly protected by the timer
reference count. The system took an extra reference on the TCP timer for the
TCP worker to prevent sockets from requeuing the timer, which crashes the
system. This, however, was not enough. There existed a race condition when
the TCP worker thread released that extra reference on the timer. If it
saw that it did not release the last reference, it would start to requeue
the timer. But before that extra reference for the TCP worker could be
taken again, sockets could decrement and then increment the reference count
from 0 to 1. This would cause the socket to start requeuing the timer as
well. With two queue attempts in the pipeline, a crash was inevitable.

The fix is to do away with that extra reference for the TCP worker and use
a separate variable instead - the TCP timer state. All threads trying to
queue the timer need to do an atomic compare exchange before proceeding.
There shouldn't be any more double queues.